rpa/requirements.txt
playwright>=1.40.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
fastapi>=0.110.0
uvicorn>=0.27.0
cryptography>=42.0.0
pydantic>=2.0.0
aiohttp>=3.9.0
